Refactor packages format
