+++ app/src/main/java/fgcu/smartclock/models/IPApiModel.kt
@@ -1,4 +1,4 @@
-package fgcu.smartclock
+package fgcu.smartclock.models
 
 import com.google.gson.annotations.SerializedName
 import java.io.Serializable
