+++ app/src/main/java/fgcu/smartclock/models/WeatherConditionModel.kt
@@ -3,9 +3,11 @@ package fgcu.smartclock.models
 import com.google.gson.annotations.SerializedName
 import java.io.Serializable
 
-data class WeatherCondition(
-  @SerializedName("main") val weatherName: String,
+data class WeatherConditionModel(
+  @SerializedName("main")
+  val weatherName: String,
   @SerializedName("description")
   val weatherDescription: String,
-  @SerializedName("icon") val weatherIcon: String
+  @SerializedName("icon")
+  val weatherIcon: String
 ) : Serializable
